pyproject.toml
[project]
name = "magic-eraser-backend"
version = "0.1.0"
description = "Magic Eraser backend using LaMa inpainting model"
requires-python = ">=3.10,<3.13"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "ultralytics>=8.3.0",
]
